misc: move IRDLDialectLoader to own file and remove circular dependency
